Add firebase project, add functionality to sign up activity, only validates form
