+++ ProjectManagement/app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.example.projectmanagement"
+        applicationId "com.ceciliadb.projectmanagement"
         minSdkVersion 22
         targetSdkVersion 29
         versionCode 1
