+++ ProjectManagement/app/src/androidTest/java/com/ceciliadb/projectmanagement/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.projectmanagement
+package com.ceciliadb.projectmanagement
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
