+++ ProjectManagement/app/src/main/java/com/ceciliadb/projectmanagement/activities/SignInActivity.kt
@@ -1,9 +1,8 @@
-package com.example.projectmanagement.activities
+package com.ceciliadb.projectmanagement.activities
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.WindowManager
-import com.example.projectmanagement.R
+import com.ceciliadb.projectmanagement.R
 import kotlinx.android.synthetic.main.activity_sign_in.*
 
 class SignInActivity : BaseActivity() {
@@ -16,6 +15,7 @@ class SignInActivity : BaseActivity() {
     }
 
     private fun setupActionBar(){
+        //Prepare the action toolbar with the icon we have.
         setSupportActionBar(toolbar_sign_in_activity)
         val actionBar = supportActionBar
         if(actionBar != null){
@@ -23,6 +23,7 @@ class SignInActivity : BaseActivity() {
             actionBar.setHomeAsUpIndicator(R.drawable.ic_arrow_back_black_24dp)
         }
 
+        //Sets the toolbar in the sign up activity (back icon) to act as if you press the back button.
         toolbar_sign_in_activity.setNavigationOnClickListener {onBackPressed()}
     }
 }
