+++ ProjectManagement/app/src/main/java/com/ceciliadb/projectmanagement/activities/SplashActivity.kt
@@ -1,12 +1,11 @@
-package com.example.projectmanagement.activities
+package com.ceciliadb.projectmanagement.activities
 
 import android.content.Intent
 import android.graphics.Typeface
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.view.WindowManager
-import com.example.projectmanagement.R
+import com.ceciliadb.projectmanagement.R
 import kotlinx.android.synthetic.main.activity_splash.*
 
 class SplashActivity : BaseActivity() {
