--- ProjectManagement/app/src/main/java/com/example/projectmanagement/activities/SignUpActivity.kt
@@ -1,28 +1,0 @@
-package com.example.projectmanagement.activities
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.WindowManager
-import com.example.projectmanagement.R
-import kotlinx.android.synthetic.main.activity_sign_up.*
-
-class SignUpActivity : BaseActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sign_up)
-        window.setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN)
-        setupActionBar()
-    }
-
-    private fun setupActionBar(){
-        setSupportActionBar(toolbar_sign_up_activity)
-        val actionBar = supportActionBar
-        if(actionBar != null){
-            actionBar.setDisplayHomeAsUpEnabled(true)
-            actionBar.setHomeAsUpIndicator(R.drawable.ic_arrow_back_black_24dp)
-        }
-
-        toolbar_sign_up_activity.setNavigationOnClickListener {onBackPressed()}
-    }
-}
